Move MockDB to new file
